requirements.txt
numpy
opencv-python
pyserial
pyrealsense2
yolov5
